requirements.txt
requests
flask
python-docx
markdown2
markdown
Flask-Markdown
